+++ components/support/source/SSynchronousScheduler.cpp
@@ -11,15 +11,15 @@
  *  @copyright (c) 2014 Draconic Entertainment
  */
 
-#include <core/SSynchronousScheduler.hpp>
+#include <support/SSynchronousScheduler.hpp>
 
 namespace Kiaro
 {
     namespace Support
     {
-        static SSynchronousScheduler *sInstance = NULL;
+        static SSynchronousScheduler* sInstance = NULL;
 
-        SSynchronousScheduler *SSynchronousScheduler::getPointer(void)
+        SSynchronousScheduler* SSynchronousScheduler::getPointer(void)
         {
             if (!sInstance)
                 sInstance = new SSynchronousScheduler;
@@ -35,9 +35,9 @@ namespace Kiaro
             sInstance = NULL;
         }
 
-        CScheduledEvent *SSynchronousScheduler::schedule(EasyDelegate::IDeferredCaller *cachedDelegate, const Common::U32 &waitTimeMS, const bool &recurring)
+        CScheduledEvent* SSynchronousScheduler::schedule(EasyDelegate::IDeferredCaller* cachedDelegate, const Common::U32& waitTimeMS, const bool& recurring)
         {
-            CScheduledEvent *event = new CScheduledEvent(cachedDelegate, waitTimeMS, recurring);
+            CScheduledEvent* event = new CScheduledEvent(cachedDelegate, waitTimeMS, recurring);
             mScheduledEventSet.insert(event);
 
             return event;
@@ -47,10 +47,10 @@ namespace Kiaro
         {
             const Common::U64 currentSimTimeMS = Support::FTime::getSimTimeMilliseconds();
 
-            for (std::set<CScheduledEvent *>::iterator iterator = mScheduledEventSet.begin(); iterator != mScheduledEventSet.end(); iterator++)
+            for (std::set<CScheduledEvent*>::iterator iterator = mScheduledEventSet.begin(); iterator != mScheduledEventSet.end(); iterator++)
             {
                 bool shouldRemoveCurrentEvent = false;
-                CScheduledEvent *currentEvent = *iterator;
+                CScheduledEvent* currentEvent = *iterator;
 
                 if (currentEvent->shouldDispatch(currentSimTimeMS))
                 {
